UHF-10565: Make RemoteEntityAccessTest extendable
