+++ src/Entity/RemoteEntityInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Drupal\helfi_api_base\Entity;
+
+use Drupal\Core\Entity\ContentEntityInterface;
+
+/**
+ * Interface for remote entities.
+ */
+interface RemoteEntityInterface extends ContentEntityInterface {
+}
